Bugfix/authentication does not always show in swagger UI (#3) (#4)
* Added limit to cRestChildCollection

You can now limit the amount of records returned by the cRestChildCollection class by setting the piLimitResult property.

* Fixed typo in documentation

POST, PUT, PATCH and DELETE methods were refering to the GET verb. Additionally described the OnHttpGetByID a little better.

* Fixed readme

Fixed links in readme, additionally fixed the import statement for the index.html not being correct.

* Fixed authentication not showing in Swagger UI on some levels

Due to a refactor the Parent object of the cOpenApiSpecification is no longer the cWebApi object. This caused certain behavior to break. Added a helper function that keeps checking parents untill it reaches the cWebApi object.