Allow psPath to be used as table name
In certain situations you do not want to use the actual table name in the swagger ui but instead use a name you can define yourself such as psPath. The new pbUsePathAsTableName allows the path to be used as a table name. This way your table names do not get exposed to the end user if you do not wish for that to happen.

Additionally changed all DF_FILE_ROOT_NAME to DF_FILE_LOGICAL_NAME to prevent the driver being appended infront of the table name.