Sync beta with stable (#22)
* Bugfix/authentication does not always show in swagger UI (#3)

* Added limit to cRestChildCollection

You can now limit the amount of records returned by the cRestChildCollection class by setting the piLimitResult property.

* Fixed typo in documentation

POST, PUT, PATCH and DELETE methods were refering to the GET verb. Additionally described the OnHttpGetByID a little better.

* Fixed readme

Fixed links in readme, additionally fixed the import statement for the index.html not being correct.

* Fixed authentication not showing in Swagger UI on some levels

Due to a refactor the Parent object of the cOpenApiSpecification is no longer the cWebApi object. This caused certain behavior to break. Added a helper function that keeps checking parents untill it reaches the cWebApi object.

* Added checks for hoKeyField (#5)

Framework now checks if hoKeyField is zero. If it is it can't find a primary key field and can't seed the buffer to perform finds.

* Allow non numeric ids in path (#6)

Previously OnHttpRquest checked if the last path of the URL was numeric and if it was marked that as the ID. Now the framework checks if the last part of the url is not equal to the sPath of the endpoint, if it isnt we can assume its a ID.

Also commited a fix with a corrupted K1 file that somehow grew to 2MB

* Fixed custom endpoints not being marked as custom (#8)

Custom endpoints were not properly being marked as custom. This caused iterators to append their own headers such as "content-type: application/json" which messed with file downloads.

* Fix certain verbs not appearing as authenticated for custom endpoints (#9)

* Allow psPath to be used as table name (#10)

In certain situations you do not want to use the actual table name in the swagger ui but instead use a name you can define yourself such as psPath. The new pbUsePathAsTableName allows the path to be used as a table name. This way your table names do not get exposed to the end user if you do not wish for that to happen.

Additionally changed all DF_FILE_ROOT_NAME to DF_FILE_LOGICAL_NAME to prevent the driver being appended infront of the table name.

* Allow readonly to work for cRestEntity and cRestChildCollection (#11)

* Allow readonly to work for cRestEntity and cRestChildCollection

You might want to hide the cRestEntity and cRestChildCollections during a post request, this allows you to do so.

* Extended the sample to showcase how to post parent records

* Fixed validation tables being grabbed from wrong field (#13)

Validation tables are now retrieved based on the server and not Main_DD. Additionally removed unused End_Construct_Objects

* Updated sample (#14)

Get endpoint for the custom endpoint is now capable of downloading the avatar of a sales person. Also added a example bad request response that shows how you can define a json object response.

* Add Note section to README

* Errors are now always shown when running in the debugger (#18)

* Errors are now always shown when running in the debugger

* Always forward error to global error handler

* Fix merge conflict (#20)

* Added limit to cRestChildCollection

You can now limit the amount of records returned by the cRestChildCollection class by setting the piLimitResult property.

* Bugfix/authentication does not always show in swagger UI (#3) (#4)

* Added limit to cRestChildCollection

You can now limit the amount of records returned by the cRestChildCollection class by setting the piLimitResult property.

* Fixed typo in documentation

POST, PUT, PATCH and DELETE methods were refering to the GET verb. Additionally described the OnHttpGetByID a little better.

* Fixed readme

Fixed links in readme, additionally fixed the import statement for the index.html not being correct.

* Fixed authentication not showing in Swagger UI on some levels

Due to a refactor the Parent object of the cOpenApiSpecification is no longer the cWebApi object. This caused certain behavior to break. Added a helper function that keeps checking parents untill it reaches the cWebApi object.

* Implement property to disable ID routing (#21)

In some instances you don't want the request to be routed to OnHttpGetByID when the last part of the path does not match the endpoint path. Especially in custom endpoints. This feature allows you to disable that behavior and always routes to OnHttpGet

---------

Co-authored-by: Marcia.Booth <Marcia.Booth@dataaccess.com>
Co-authored-by: DataFlex.dev <127292730+DataFlex-dev@users.noreply.github.com>